Add option for label-based augmentations
Implemented a new command-line argument to enable Random Labels To Image augmentation in MRI data preprocessing. This feature, which defaults to off, allows users to opt-in for an additional augmentation method, aiming to further enhance the dataset diversity for training machine learning models. Now, the actual augmentation selection respects this new setting, ensuring augmentations are applied only when enabled.