Refactor ndimage imports; add levels extraction script
Refactored import statements to use alias 'ndi' for scipy.ndimage across multiple utility scripts for consistency and readability. Added a new command to the inference script for extracting vertebrae and IVD levels. Also updated the argparse documentation of the extract_soft script to provide a clear description of its functionality.