Add mapping of input labels to output segmentation
Introduced a new argument for providing a mapping of input labels to
output segmentation labels. This facilitates custom label conversions,
such as mapping the input C1 label to a specified output label.

Updated related functions and modules to handle and apply this mapping
throughout the segmentation process. This enhancement increases
flexibility and accuracy in label assignments.