Refactor landmark output label determination strategy prioritizing localizer
Refactored the logic for setting initial landmark output labels by prioritizing localizer data first and then segmentation data if necessary. This allows for a more systematic approach, increasing reliability in label determination. Additionally, moved the fallback check for default superior label to the end of the process, ensuring it's only used as a last resort.